update_tools: update_template: Added custom boot that ct will run
CT wants to the system state to stay in updating after hmupdate.img is
finished. It is because they are piggy back riding it and will install
i4m application once we have booted linux.

So created custom ctubiboot and removed sppi cmd's from ubiboot. ubiboot
is the one called inside the hmupdate.img and ctubiboot will be called
in normal start up.