u-boot-t20: board: toradex: Added method to ping MX-4 co-cpu
During the boot sequence we tell the co-cpu in which stage we are. The
co-cpu acts as an watchdog.

On some hardware we need a first ping within a few seconds to tell the
co-cpu that we have started up.

Added method to ping the co-cpu in board_late_init. This is controlled
by the CONFIG_PING_MX4 define.