u-boot-t20: configs: Added PING_MX4 option to all mx4 configs
Before releasing a pic firmware which requires a PING within 5 seconds
we enable it in all u-boot configurations.

If co-cpu is not pinged within 5 seconds it will see that as an failed
boot attempt and reset by shutting down 3.3-PERM (both main and co-cpu
runs on this voltage). This is first valid in 1.2.1 co-cpu firmware.