u-boot: toradex: board: Fixed up ping_mx4_pic
We where causing problems with the spi bus with our ping command.
Following command was out of sync, meaning that update blink will never
be triggered.

What was needed is a release of the bus between the two commands. Also
added a 10 ms delay between the commands just to make sure we are not
racing here.