update_template: Add a small delay before starting the USB probe
When probing USB for an hmupdate.img we need to insert a small delay.
This is mostly for the OTG port where have som logic in co-cpu to check
if we have a device connected or if we are a host. We need to wait for
this probe to finish before we can start the USB in u-boot. Otherwise we
could miss to detect the USB storage device and hence fail to load
hmupdate.img from USB when inserted to OTG port.