cmd_source: HACK: set fit image length to -1
If image length is set we will try and allocate that amount of memory from
the malloc pool and then perform a copy. The malloc pool is normally
small (few MB) but our images can be > 100 MB.

So to not increase the malloc pool to large amounts we set image size to -1
which means that it will re-use the address that is passed to the source command.

This is probably a bug, and might we a "proper" solution to this when time allows it.

Signed-off-by: Mirza Krak <mirza.krak@hostmobility.com>