mtd: tegra_nad: sync with Linux driver
The NAND driver used in Linux and U-boot need to agree on how to
access the chip and where OOB is and length of it. And also ECC
algorithm and so on.

As we have updated Linux kernel to 4.1 which has an updated NAND
driver we need to syncronize the driver in U-boot for it to work
properly

Signed-off-by: Mirza Krak <mirza.krak@endian.se>